+++ src/RestApi/Policy/IPolicyFilter.ts
@@ -1,0 +1,5 @@
+export default interface IPolicyFilter {
+	pagination?: number;
+	createdUntil?: Date;
+	archived?: boolean;
+}